make create_page_aad method of CryptoContext
